seperated class objects from functions
